requirements.txt
djangorestframework
markdown
django-filter
python-decouple